FastAPI/requirements.txt
fastapi==0.115.0
python-dotenv
openai==1.51.2
pydantic
pillow